I HAVE SOME OF THE BASIC VIEWS IN PLACE
I added some basic styling such as the color theme from
_variables.scss to index.scss, and set the font-size to 16 px
in order to establish rem units.

I added “testprep” to the header control brand, and added the select
test element to go back to the select test view.

I overrode the boostrap primary color in _variables.scss to black.

I added the HomeView, SelectTestView, and TakeTestView routes to
App.jsx.

I altered SignIn.js so that once signed in, you are directed to
SelectTestView.

I added the initial data-model.js file which returns all tests and a
specific test by test ID.

I added the views folder, and created shells for HomeView,
SelectTestView, and TakeTestView. SelectTestView.jsx shows some basic
cards with a button to select a test and it navigates to the
TakeTestView.