I HAVE A MINMALLY VIABLE TEST
The data-model.js file was moved to the api folder where it should be.

I added a Question.jsx component that know how to render a question.

I added the file questionTypeEnum that defines the types of questions
to render in the new question.jsx component.

TakeTestView was modified to use the new Question component.
It also provides a callback for passing back the answers as people
select and unselect radio buttons and checkboxes. It also now has a
scoring algorithm.

This is good enough for an MVP. I shall comeback towards the end and
provide more styling an functionality such as showing your score. I'll
come back to this towards the end.