I ADDED THE VIEWS FOR EDITING TESTS
In App.jsx I added routes for MyTests and EdiTests. The former allows
the user to select a test to edit and the latter allows for CRUD.

In Header.jsx, I changed the name of Select Test to Take a Test.
I also added the link for My Tests to select a test to edit.

In SelectTestView.jsx, I changed the text “Select A Test" to
"Take A Test".

I added the views that let you select a test to edit, and then an empty
view where you actually edit a test.