MOST OF THE CONTROLS ARE IN PLACE FOR EDIT TEST
Most of the controls are in place, but not working. I am limiting the
tests to 4 choices. THe back and next question buttons are in place but
not working at the moment. There are also Add and Delete question
buttons but are not working at the moment.

The text fields are bound to state.

The select best and select all that apply radio buttons force the
is correct check fields to render as radio buttons or check boxes. This
is actually working at the moment.

I am making a commit so that I don't lose any work.