MORE CHANGES TO EDIT TEST
There is now a method “setQuestionState” the sets the state for the
currently displayed question. This reduces a lot of duplicate code.

The isCorrect buttons are now disabled if their corresponding text
field are empty.

The previous and next buttons now enabled and disabled as appropriate
and now load the current question.

You can now add a new question to the test.

The radio and checkboxes for the choices now are selected and
deselected as appropriate.