CLOSER TO A WORKING EDIT TEST VIEW
SetQuestionState has small tweaks to it.

I added the saveTheCurrentQuestionState method that takes React’s
flattened state and overwrites the fields in the JS object that we send
and receive from the server. This method is invoked whenever we
navigate to a different question, or we delete or add a new question.

The nav buttons now enable and disable properly.