YOU CAN NOW ADD A NEW EMPTY TEST
I Added the createATest method to data-model.js in order to create a
new test.

I added a card to MyTestsView.jsx which is shown first. It is a card
that had a button that lets you create a new test.

EditTestView.jsx now has the code to create a new test. If you create a
new empty test, then the POST route is selected; otherwise the PATCH
route is selected.