DELETE A TEST IS NOW WORKING; FE FULLY FUNCTIONAL
In data-model.jsx the deleteATest method is working.

All of the delete functionality has been added to MyTests.jsx. The
cards for editable tests now have a delete button. The delete
functionality removes the test from the array of tests which is also
duplicated in state.

The frontend is now fully functional and ready for a test deployment.
Of course it needs TLC to be production ready.