TAKETESTVIEW.JSX NOW KEEPS SCORE
I have an unstyled span that keeps score. I also display the results
at the end of the test in the alert. This is good enough for now. At
least the user is notified.

Working on this, I thought that I could leave the radio buttons
uncontrolled and try using Refs.