EDITESTVIEW NOW PREVENTS SAVING AN INVALID TEST
I had the problem that when selecting a test to take, there could be
a test with no questions and a test with questions but no choices.

Now I validate a test and a question before allowing you to save it.

But this now saves me work. I am now seeing light at the end of the
tunnel.